Refactoring to make linux tests pass
